refactor(aigateway): parse float params as *float64
JSON decodes straight to float64 with no precision noise, so the
float32To64 widening helper was solving a problem the float32 fields
themselves created.

+++ pkg/util/aigateway/aigateway.go
@@ -17,8 +17,8 @@ type ParsedInferenceRequest struct {
 	URL                 string   `json:"url"`
 	Model               string   `json:"model"`
 	Seed                *int     `json:"seed,omitempty"`
-	Temperature         *float32 `json:"temperature,omitempty"`
-	TopP                *float32 `json:"top_p,omitempty"`
+	Temperature         *float64 `json:"temperature,omitempty"`
+	TopP                *float64 `json:"top_p,omitempty"`
 	MaxTokens           int      `json:"max_tokens,omitempty"`
 	MaxCompletionTokens int      `json:"max_completion_tokens,omitempty"`
 	StopSequences       []string `json:"stop,omitempty"`
@@ -93,11 +93,11 @@ func ParseInput(req Request) (ParsedInferenceRequest, error) {
 			return ParsedInferenceRequest{}, err
 		}
 
-		// rf uses non-pointer floats, so params where an explicit zero is
+		// rf uses non-pointer float32s, so params where an explicit zero is
 		// meaningful (e.g. temperature 0) must be re-parsed as pointers.
 		var floatParams struct {
-			Temperature *float32 `json:"temperature"`
-			TopP        *float32 `json:"top_p"`
+			Temperature *float64 `json:"temperature"`
+			TopP        *float64 `json:"top_p"`
 		}
 		_ = json.Unmarshal(req.Body, &floatParams)
